CA-110446: XenCenter Installation: multiple install on same OS when selected installation for all users.
Changed XeCenter.wxs file (and updated XenCenter.l10n.diff accordingly) to
set ALLUSERS property when InstallDir key found in HKLM registry, i.e.
previous version of XenCenter was installed for "All users" (the
property is not set when previous version was installed for "Just me"). The
property is cleared after searching for related products is finished to
allow the user changing installation context in GUI.

Signed-off-by: Adrian Jachacy <Adrian.Jachacy@citrix.com>