CA-108512: Fixed the issue when Connect to Server dialog box becomes unresponsive when invalid password is tried during reconnect to host.
The error occurred when user specified a second wrong password. If the same wrong password was specified again and again, it didn't trigger it. This was due to an error in code where we didn't use the updated password in the second connection attempt.
Fixed by changing the code to use the updated username and password combination for every login attempt.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>