CA-115742: On VM properties page, if a VM has no affinity, XenCenter shouldn't automatically select one.
This fix corrects the following behaviour: when you have a VM with no affinity and you go into its Properties dialog,
XenCenter believes that a VM on a standalone host has to have that affinity, and greys out the "no affinity" radio button,
which has the effect of forcibly changing the affinity to the host, which a VM Admin is not allowed to do.

The correct behaviour is that if a VM already has no affinity, we don't grey out the radio button.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>