CA-132066 Make Heartbeat work as designed
The Heartbeat would log out the session after the first missed heartbeat.
This worked fine when the Heartbeat was on its own session, back in George
and earlier versions. However, when it was changed to use just a
DuplicateSession() instead of a new Session(), this broke the main session
of the connection. This caused a first heartbeat failure to break the
connection, and also caused the expectDisruption flag to be ignored.

This patch restores the old behaviour by not logging out the session until
the second failure.