CA-136288 Restore application settings after installing a new version of XenCenter
When running XenCenter after a new installation, and the program's hash has changed (e.g. by .NET 4.0 upgrade), the previous settings are not automatically restored because they cannot be found (they are there, but in different folder).
We need to try and locate a config file from a previous installation, and if found, update the settings from there.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>