CP-8547: Add button in XenCenter to initiate space reclamation on an SR.
The operation applies to an SR. The button is on the Storage tab of the host/pool, and is entitled "Reclaim freed space".
It comes between the New SR and Properties buttons.
If the SR does not support TRIM , the button is invisible and the Properties button moves left to close up the gap.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>