CA-149811: End the connection to each slave after that host has joined the pool to prevent XenCenter showing connection lost alerts
This fixes the following issue: event.from fails with "The underlying connection was closed: A connection that was expected to be kept alive was closed by the server.".
It happens because we end the connections to the slave hosts only after all hosts have joined the pool and in the meantime event.from on a slave fails.

We fix this by ending the connection to each host immediately after the host has joined the pool.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>