CA-184560: Fix the issue when the RPU wizard cannot check the host version after reboot
In the RPU wizard, after we notice that the host has been rebooted we try to resolve the host from cache, but sometimes the cache is not populated at that point.
Then, because we cannot resolve the host, we cannot check the host version and we assume is a newer version (which might not be true).
This commit adds a call to retry to resolve the host (with timeout).

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>