CA-189674: Move VM shows an irrelevant (VIF mapping) error, but it shouldn't
I assumed the problem was that vm.resident_on is always unknown for stopped vms, but this is there intentionally to match only running vms and be always false for stopped ones. However, we still need similar logic to not to send VIF mapping when not needed in the assert.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>