CA-196231: Don't confuse start-on with migration
The "Start on Server" VM context menu incorrectly enabled hosts that a VM can
only be started on after a storage migration.  This was due to the fact that
the VMOperationToolStripMenuItem class is used for both "Start on Server" as
well as "Migrate to Server".

This change ensures that start checks are done when a start is requested, and
migration checks otherwise.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>