CA-199145: Home page missing the page icons that are used as list bullets
Now the Mhtml file contains this image as well. I used IE to save it and apparently (like other browsers) it does not save css-specified background images, so had to add a 0x0px sized img into the html to have IE to save it. Easy workaround and no change to the layout.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>