CA-202374: XC greyed out 'Install Tools' menu for imported VM from Cream
Check different feature flag for Windows VMs on Dundee to detect Dundee guest agent. For earlier versions nothing is changed.
This will cause XC to report "Management not installed" for Windows VMs on Dundee hosts with Cream guest agent (This is the case when a host has been upgraded from Cream to Dundee).

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>