CA-210732: XenCenter VM properties GPU always shows "None"
The error occurs when there is a gpu group with only one vgpu type: When setting the selected item in the vGPU combobox, XenCenter tries to select the group header instead of the vGPU type; because the group header is not selectable, we then default to "None".
With this fix, we ensure that two items are not equal if one is a group header and the other is a subitem.

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>