CP-15714: UI: Change Select Servers Page
DownloadUpdatesXmlAction is called on PageLeave and not when the user selects Automatic Mode. This way we get the error message at the right time, also showing a progress dialog what makes the UX much better.
New tooltips for servers on the Select Server Page to to cover two more cases

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>