[CA-211369] Console switching when snapshot completes
ConsolePanel.setCurrentSource() being called in ConsolePanel.Snapshot() was causing the console view to switch to the target snapshotted VM's console, but only if the user is already on a console tab. The screenshot now happens before the snapshot begins, instead of after; this way, the screenshot always occurs when not on the console tab.

Signed-off-by: Frederico Mazzone <fredericom@citrite.net>