CP-18013: [XenCenter] Update host updates pane
For servers running Ely or greater we now produce messages for patching requiring reboot using the host.patches_requiring_reboot field. The Pool_patch objects resolved are sufficient to produce the same messages. Note that some patches produce a message in the old API but not the new, this is intentional. Still to do: update pool updates in the same way,

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>