CP-18014: Alter batch hotfixing to use host.patches_requiring_reboot
XenCenter will check host.patches_requiring_reboot when a host restart is due while a batch update is in progress. If the restart is not required, XenCenter will skip relevant PlanActions.
The intention here is to check this whenever a restart is due according to the original schedule - this way a completely livepatchable update can prevent the restart while the key batch hotfixing logic is not changed.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>