[CP-19111] Added ability to choose Basic or Digest authentication for proxy server
2 new radio buttons in Connection Options page, for Basic and Digest; Digest is the default as it's the most secure.
New proxy authentication method setting, which is used to configure XenAPI's HTTP class and the .NET AuthenticationManager (which handles which authentication schemes can be used by the .NET web classes).
The new setting is also transferred and used by the Health Check service.

The bug noted in CA-214653 also occurs here, but the work-around on PR#1201 for that bug will also work for this.

Signed-off-by: Frezzle <frederico.mazzone@citrix.com>