[CA-220457] Add friendly names for new graphs, and filter duplicates
We have some new RRDs - read, write, read_latency and write_latency. This adds support for them by adding a friendly name and using it to display them. It also adds some filtering (see GraphHelpers.cs) because for backward compatibility io_throughput_read and io_throughput_write (which differ from read and write only in scale) are still available, but shouldn't be displayed in XC. If the new read or write are available, we filter these old datasources out.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>