CA-227067: General tab: XenCenter does not show Reboot required when Toolstack restart required
On the General tab XenCenter shows a list of updates that requires the host to be restarted. This is shown for applied updates for which the required guidance hasn't been done (eg. restartHost or restartAgent after-apply-guidances)

This commit fixes a regression that caused restartToolstack warnings to be not shown in the list of warnings.

The displayed messages are also improved, XenCenter will show Toolstack restart or host restart appropriately and not just restart only as it used to.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>