[CA-226897] Resume on Server option won't automatically resume the VM
The resume on server uses the CrossPoolMigrateCommand to implement the migration of the VM to the desired host. Before this change that command had no support for resuming a VM after migrating it, so the VM was not resumed. With this change, the CrossPoolMigrateWizard can take an optional (default false) parameter to restart the VM after migration. When this is true, the migrate action becomes a MultipleAction, first migrating and then restarting the VM.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>