[CA-228051] XenCenter allows user to create unusable PVS Proxy cache config
A site's configuration needs to be completed outside of XenCenter, and when it isn't done we shouldn't allow VMs to enable PVS read caching on that site. We now disable sites that aren't configured in the enable read caching dialog, using an EnableableComboBox. I've also changed the DrawMode of that ComboBox component, OwnerDrawFixed more appropriately sizes the dropdown container when the number of items is small.