CA-232294: Can't connect if an Add Server dialog has been cancelled before
This commit is to fix the following: "when right clicking a disconnected pool the context menu gives you the option to connect, this opens a window to enter credentials. if you close this window (either by cancel or the X), next time when you right click and click connect, nothing happens. you have to restart XenCenter to connect to the pool"

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>