[CA-233454] PVS tab doesn't show a new VM when it's created (#1314)
* [CA-233454] PVS tab doesn't show a new VM when it's created

Refined the rules for not adding a VM to the table, if it is a template (thus not_a_real_vm), and it has the __gui__ prefix (thus hidden), we still add it, but hide it.
When its name is changed (to remove the __gui__ prefix), we update its name and re-calculate whether it should be visible (in the case of a new VM this will be true once the __gui__ prefix is gone). Also resort the table if a node changes from hidden to visible, because it appears as an addition to the table.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>

* [CA-233454] Update criteria for VM visibility to include is_a_real_vm

is_a_template is changed before the name_label removes the __gui__ prefix, so this works with no other changes to the vm property changed event.

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>

* [CA-233454] Update when the VMs are shown in the PVS list

New observable property IsBeingCreated for VMs, set to true when they're made a hidden object, and false when they're removed from hidden objects (both in CreateVMAction). In the PVS Page when this is set to false, we re-evaluate whether a VM can be shown. This means that new VMs show here at the same time they're added to the tree (only different is tree refresh time), instead of far earlier (and before their networks were added).

Signed-off-by: Callum McIntyre <callumiandavid.mcintyre@citrix.com>

* [CA-233454] Set IsbeingCreated in the CreateVMFastAction

* [CA-233454] Properly support the VM Fast Create action

Further changes to CreateVMFastAction, to ensure it works with PVS tab - use the __gui__ prefix when the VM is created and then change it back just before showing.

* [CA-233454] Small logic adjustments/tidying up