[CP-19888] Changes to service stop/start
Always stop any running ncover service before starting (instead of ignoring the error and using the currently running one), and ignore the error if it's already been stopped at the end. The logic for the latter is that we stop the service right after the commands that require it, so if it stopped too early they'll error anyway. If they don't error because of the stopped service then we shouldn't error right at the end.