CA-226900: Do not allow selection of individual pool members when installing a hotfix
This change does not apply to supplemental packs, as we should be able to install supp packs to individual hosts in a pool.
Also, if the wizard is installing an update from disk and we cannot match it in updates.xml, then we will continue to allow host selection, because we can't tell whether it is a supp pack or a hotfix.

Plus fixing the bug where the "Select All" button is enabled even when all the selectable servers/pools are selected

Signed-off-by: Mihaela Stoica <mihaela.stoica@citrix.com>