CA-241431: Using standard .net WebClient to download updates.xml
Using standard .NET Framework functionality instead of the custom implementation of the HTTP stack. (To download a single file.)
Note that this .net class WebClient is already being used to download hotfixes. However, in that code the proxy is not set, I am fixing that in a new commit on the same ticket.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>