CA-220047: Show proper precheck error when a VM cannot be migrated
A bit of refactoring as requested by the reviewers

I decided not remove the NullReferenceExceptions and throw the Failures
immediately in order to avoid the additional error handling that seems to be unnecessary there. However, the reason I've decided to keep it as is is that it would change what gets logged and the performance gain would not be noticeable anyway. Furthermore, the current implementation (try-catchall) catches any other exceptions providing fallback to the default Failure.