CP-22787: Backporting CA-249084 follow-up commit
FalconOrGreater hasn't been implemented in Honolulu, nor
HonoluluOrGreater. We can't implement the latter as the platform_version
is not expected to change between Ely and Honolulu. To be able to backport
these changes (from PR #1536 CA-249084), the consensus reached was to use
ElyOrGreater in Honolulu. This would make the PoolJoinRules less strict
(fail later) and would hide the build number for Ely, but this seems to be the minimal impact and straightforward solution.

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>