CA-252924: Updates on new version are not applied for licensed host on Partially licensed pool
Changed the text (on the checkbox) to make it clear that, after a version upgrade, XenCenter will apply updates only on licensed pools. This way it should be clear that servers of a partially licensed pool will not be updated. (The previous text suggested that any licensed server would be updated in general.)

Signed-off-by: Gabor Apati-Nagy <gabor.apati-nagy@citrix.com>