[CA-269110] Add the VM wizard button before enabling dropdown items
We always add the VM wizard button to the migrate dropdown, but it's
currently not added until the end of the enable loop - after we've
calculated whether we can boot on each host. There's no reason to wait
until we've done all that before showing the wizard, so add it straight
away to the end of the list before the loop.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>