[CA-272150] Use async load for dropdown items
The ticket suggests re-using the calculation of whether a pool can be
enabled for whether each host can be enabled, but we don't seem to have
an easy way of caching that data. Also the logic isn't quite the same -
we can enable the pool option if the VM can migrate to any host, we
don't need to precompute them all at that point. So I don't think it'd
be a good idea to add that caching or a requirement to precompute all
hosts in order to enable a pool in the inital combobox.

Instead I've replaced the `LoadAndWait` call when items are added to the
dropdown with `Load`, which is asynchronous and doesn't calculate on the
UI thread. This gives a much smoother UI as each host is enabled
independently.

The rest of the changes are to allow the `ReasonUpdated` on each item to
call the method that enables/disables the next button - ensuring the
next button gets enabled when a host is selected in the dropdown (even
if it wasn't selectable when it was first added, because we change that
async now).

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>